+++ AudioRacer-PlayerCommunication/src/at/fhv/audioracer/communication/player/message/FreeCarsMessage.java
@@ -2,7 +2,7 @@ package at.fhv.audioracer.communication.player.message;
 
 public class FreeCarsMessage extends PlayerMessage {
 	
-	public int[] freeCars;
+	public byte[] freeCars;
 	
 	public FreeCarsMessage() {
 		super(PlayerMessage.MessageId.UPDATE_FREE_CARS);
