+++ AudioRacer-WorldSimulator/src/at/fhv/audioracer/simulator/world/SimulationController.java
@@ -49,7 +49,7 @@ public class SimulationController {
 	
 	private MapComponent _map;
 	
-	private int _carId;
+	private byte _carId;
 	
 	private List<ICarClient> _carClients;
